Rename sort file.
